requirements.txt
behave==1.2.6
Appium-Python-Client==2.11.1
pytest==7.3.2
pytest-html==3.1.1
selenium==4.10.0
requests==2.31.0
PyYAML==6.0
python-dotenv==1.0.0
allure-behave==2.13.2
pytest-xdist==3.3.1 